requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.103.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.28.0
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyngrok==6.1.0
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
twilio==8.8.0
typing_extensions==4.8.0
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
